Add timestamp injection script
