+++ _indexer/package.json
@@ -5,13 +5,17 @@
   "main": "indexer.ts",
   "type": "module",
   "scripts": {
-    "indexer": "node --loader ts-node/esm indexer.ts"
+    "indexer": "node --loader ts-node/esm indexer.ts",
+    "timestamper": "node --loader ts-node/esm timestamper.ts"
   },
   "keywords": [],
   "author": "xp",
   "license": "GPL v3",
   "devDependencies": {
-    "typescript": "^4.9.5",
-    "ts-node": "^10.9.1"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
+  },
+  "dependencies": {
+    "simple-git": "^3.27.0"
   }
 }
